blender external material browser proof-of-concept
no browser as of yet, but this is proof that blender can do an http
call, fetch .blend path and material name, link against the .blend file,
and assign the material to selected polygons

the rest is just hard work :) in general I need a way to roam through
.blend files and export materials (this is doable). then I need an
html-based material browser that lets me select a material. from there I
can use the blender addon to fetch the selected material and assign in.